extract onboarding wizard-setup logic into onboarding-setup.ts
